# internal/httpapi/openapi.yaml
openapi: 3.0.3
info:
  title: OpenComfy
  description: |
    OpenAI-compatible image/video API in front of ComfyUI.

    1. Click **Authorize** and paste `sk-...` (or use the `X-Api-Key` header).
    2. Import saved ComfyUI workflows at `/import` (multi-select). Blocked if prompt or required reference is missing.
    3. `GET /v1/models` to see configured workflows.
    4. `POST /v1/images/generations` is synchronous (may take minutes).
    5. `POST /v1/videos` returns `{status: queued}`; poll `GET /v1/videos/{id}` until `completed` (includes `url`).
  version: 0.1.0
servers:
  - url: /
    description: this server
tags:
  - name: meta
  - name: models
  - name: images
  - name: videos
  - name: chat
  - name: import
  - name: mcp
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
  /health:
    get:
      tags: [meta]
      summary: Liveness
      security: []
      responses:
        "200":
          description: ok
  /ready:
    get:
      tags: [meta]
      summary: ComfyUI reachable
      security: []
      responses:
        "200":
          description: ready
        "503":
          description: ComfyUI unreachable
  /openapi.json:
    get:
      tags: [meta]
      summary: OpenAPI document
      security: []
      responses:
        "200":
          description: spec
  /import:
    get:
      tags: [import]
      summary: HTML UI to pick ComfyUI workflows
      security: []
      responses:
        "200":
          description: page
  /v1/comfy/workflows:
    get:
      tags: [import]
      summary: List ComfyUI userdata workflows with inferred mappings
      responses:
        "200":
          description: each item has can_import, errors, params
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/comfy/analyze:
    post:
      tags: [import]
      summary: Analyze a workflow JSON or a userdata name for missing nodes and models
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                workflow:
                  type: object
      responses:
        "200":
          description: missing_nodes, missing_models, reused_models, allowlisted_packs, ready
  /v1/comfy/provision:
    post:
      tags: [import]
      summary: Save JSON to ComfyUI userdata and optionally git-clone allowlisted custom nodes
      description: |
        Variant C. Custom nodes clone only if `comfyui.custom_nodes_dir` is set and the pack
        is on the allowlist. Missing weights: search `comfyui.models_dir` by filename first
        (symlink into the expected Comfy folder); with `install_models` download the rest
        from Hugging Face. Does not reboot ComfyUI.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                overwrite:
                  type: boolean
                save:
                  type: boolean
                install_nodes:
                  type: boolean
                install_models:
                  type: boolean
                workflow:
                  type: object
      responses:
        "200":
          description: saved path, analysis, installed git dirs, needs_reboot, download_id
  /v1/comfy/downloads/{id}:
    get:
      tags: [import]
      summary: Weight download job (local reuse, then Hugging Face)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: status and per-file progress
        "404":
          description: unknown job
  /v1/comfy/import:
    post:
      tags: [import]
      summary: Import selected workflow filenames (multi)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [names]
              properties:
                names:
                  type: array
                  items:
                    type: string
                  example:
                    - MiniMax H3 Talking Avatar.json
                    - ND_Flux2_Klein_AIO_v1.json
      responses:
        "200":
          description: imported / skipped / failed
        "400":
          description: none imported (prompt/reference missing or empty selection)
  /v1/comfy/unload:
    post:
      tags: [import]
      summary: Remove models from OpenComfy API (not ComfyUI library)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items:
                    type: string
                  example: [nd-flux2-klein-aio-v1]
      responses:
        "200":
          description: removed / failed
  /mcp:
    post:
      tags: [mcp]
      summary: MCP JSON-RPC (initialize, tools/list, tools/call)
      description: |
        Streamable HTTP. Same Bearer as the REST API.
        tools/list exposes list_models, get_model, get_video, and one generate_<id> tool per catalog model
        whose inputSchema required fields match the workflow (e.g. input_image as a data URL).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: JSON-RPC result
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      tags: [mcp]
      summary: MCP GET is not used (no SSE notifications)
      responses:
        "405":
          description: method not allowed
  /v1/models:
    get:
      tags: [models]
      summary: List models
      responses:
        "200":
          description: OpenAI-style list plus parameters, required_parameters, input_schema, mcp_tool
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/models/{id}:
    get:
      tags: [models]
      summary: Get one model
      parameters:
        - $ref: "#/components/parameters/ModelId"
      responses:
        "200":
          description: model
        "404":
          description: not found
  /v1/images/models:
    get:
      tags: [models]
      summary: Image model capabilities
      responses:
        "200":
          description: OpenRouter-shaped image catalog
  /v1/videos/models:
    get:
      tags: [models]
      summary: Video model capabilities
      responses:
        "200":
          description: OpenRouter-shaped video catalog
  /v1/images/generations:
    post:
      tags: [images]
      summary: Generate image (sync)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageRequest"
            examples:
              flux:
                value:
                  model: toy-image
                  prompt: a red cube on white, studio light
                  size: 1024x1024
                  n: 1
                  response_format: b64_json
      responses:
        "200":
          description: created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageResponse"
        "400":
          $ref: "#/components/responses/APIError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/videos:
    get:
      tags: [videos]
      summary: List this key's video jobs
      responses:
        "200":
          description: list
    post:
      tags: [videos]
      summary: Create video job (async, always 200 queued)
      parameters:
        - in: query
          name: wait
          schema:
            type: boolean
          description: long-poll until completed/failed (capped by max_wait_s)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoRequest"
            examples:
              t2v:
                value:
                  model: minimax-hailuo-02
                  prompt: ocean waves at sunset, cinematic
                  seconds: "6"
          multipart/form-data:
            schema:
              type: object
              properties:
                model:
                  type: string
                prompt:
                  type: string
                seconds:
                  type: string
                input_reference:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: one or more files; repeat the field
      responses:
        "200":
          description: queued job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoJob"
  /v1/videos/{id}:
    get:
      tags: [videos]
      summary: Job status (completed includes url)
      parameters:
        - $ref: "#/components/parameters/VideoId"
      responses:
        "200":
          description: job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoJob"
        "404":
          description: not found
    delete:
      tags: [videos]
      summary: Cancel job
      parameters:
        - $ref: "#/components/parameters/VideoId"
      responses:
        "200":
          description: cancelled
  /v1/videos/{id}/content:
    get:
      tags: [videos]
      summary: Download bytes (OpenAI SDK)
      parameters:
        - $ref: "#/components/parameters/VideoId"
      responses:
        "200":
          description: video bytes
          content:
            video/mp4:
              schema:
                type: string
                format: binary
  /v1/files/{id}:
    get:
      tags: [videos]
      summary: HMAC or bearer file download
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: exp
          schema:
            type: integer
        - in: query
          name: sig
          schema:
            type: string
      responses:
        "200":
          description: bytes
  /v1/chat/completions:
    post:
      tags: [chat]
      summary: Image chat shim
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                model:
                  type: string
                  example: toy-image
                messages:
                  type: array
                  items:
                    type: object
                    properties:
                      role:
                        type: string
                      content:
                        type: string
            example:
              model: toy-image
              messages:
                - role: user
                  content: a red cube
      responses:
        "200":
          description: chat.completion with images[] and data[]
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: sk-
      description: OpenAI-style API key
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
  parameters:
    ModelId:
      in: path
      name: id
      required: true
      schema:
        type: string
    VideoId:
      in: path
      name: id
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: invalid api key
    RateLimited:
      description: RPM or GPU busy
    APIError:
      description: OpenAI-style error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            type:
              type: string
            code:
              type: string
            param:
              type: string
              nullable: true
    ImageRequest:
      type: object
      required: [prompt]
      properties:
        model:
          type: string
        prompt:
          type: string
        negative_prompt:
          type: string
        n:
          type: integer
          default: 1
        size:
          type: string
          example: 1024x1024
        width:
          type: integer
        height:
          type: integer
        seed:
          type: integer
        quality:
          type: string
        response_format:
          type: string
          enum: [b64_json, url]
        input_image:
          type: string
          description: data URL, or use input_images for several
        input_images:
          type: array
          items:
            type: string
          description: extra reference images (data URLs), mapped in order to LoadImage nodes
    ImageResponse:
      type: object
      properties:
        created:
          type: integer
        data:
          type: array
          items:
            type: object
            properties:
              b64_json:
                type: string
              url:
                type: string
              revised_prompt:
                nullable: true
    VideoRequest:
      type: object
      required: [prompt]
      properties:
        model:
          type: string
        prompt:
          type: string
        seconds:
          description: duration, string or number
          example: "6"
        duration:
          description: alias of seconds
        seed:
          type: integer
        input_reference:
          description: "data URL, {image_url: data:...}, file, or an array of those"
        input_images:
          type: array
          items:
            type: string
          description: additional data URLs; order matches maps_to LoadImage nodes
    VideoJob:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
          example: video
        model:
          type: string
        status:
          type: string
          enum: [queued, in_progress, completed, failed, cancelled, expired]
        progress:
          type: integer
        queue_ahead:
          type: integer
          description: generations that run before this job; 0 when it is already running or finished
        url:
          type: string
          description: HMAC file URL; required when status=completed
        unsigned_urls:
          type: array
          items:
            type: string
        created_at:
          type: integer
        completed_at:
          type: integer
          nullable: true
        error:
          nullable: true